arm: Explicitly check for sys/auxv.h and getauxval()
As reported on the issue tracker, the getauxval() interface is actually
not specific to glibc. It is also implemented in musl libc,
and apparently also Android has it.

Since the auxvals are provided by the kernel, loosen the check to just
Linux, combined with autoconf checks for sys/auxv.h + getauxval().

Bug: https://github.com/mgorny/cpuid2cpuflags/issues/3